# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Crypto Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /strategies:
    get:
      operationId: listStrategies
      tags: [strategies]
      summary: List all strategies
      responses:
        "200":
          description: List of strategies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyList"
    post:
      operationId: createStrategy
      tags: [strategies]
      summary: Create a new strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "201":
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{strategyId}:
    get:
      operationId: getStrategy
      tags: [strategies]
      summary: Get strategy details
      parameters:
        - in: path
          name: strategyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
    put:
      operationId: updateStrategy
      tags: [strategies]
      summary: Update strategy config (live)
      parameters:
        - in: path
          name: strategyId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyUpdate"
      responses:
        "200":
          description: Updated strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
    delete:
      operationId: deleteStrategy
      tags: [strategies]
      summary: Delete a strategy
      parameters:
        - in: path
          name: strategyId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /strategies/{strategyId}/activate:
    post:
      operationId: activateStrategy
      tags: [strategies]
      summary: Activate a strategy
      parameters:
        - in: path
          name: strategyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Strategy activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{strategyId}/deactivate:
    post:
      operationId: deactivateStrategy
      tags: [strategies]
      summary: Deactivate a strategy
      parameters:
        - in: path
          name: strategyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Strategy deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"

  # ── EXCHANGES ─────────────────────────────────────────────────────────────────
  /exchanges:
    get:
      operationId: listExchanges
      tags: [exchanges]
      summary: List connected exchanges
      responses:
        "200":
          description: List of exchanges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeList"
    post:
      operationId: addExchange
      tags: [exchanges]
      summary: Add an exchange connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeInput"
      responses:
        "201":
          description: Exchange added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exchange"
  /exchanges/{exchangeId}:
    get:
      operationId: getExchange
      tags: [exchanges]
      summary: Get exchange details
      parameters:
        - in: path
          name: exchangeId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exchange details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exchange"
    delete:
      operationId: removeExchange
      tags: [exchanges]
      summary: Remove an exchange connection
      parameters:
        - in: path
          name: exchangeId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
  /exchanges/{exchangeId}/test:
    post:
      operationId: testExchange
      tags: [exchanges]
      summary: Test exchange connectivity
      parameters:
        - in: path
          name: exchangeId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeTestResult"
  /exchanges/{exchangeId}/balance:
    get:
      operationId: getExchangeBalance
      tags: [exchanges]
      summary: Get exchange account balances
      parameters:
        - in: path
          name: exchangeId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exchange balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeBalance"
...
  /backtests:
    get:
      operationId: listBacktests
      tags: [backtests]
      summary: List backtest runs
      responses:
        "200":
          description: List of backtests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestList"
    post:
      operationId: runBacktest
      tags: [backtests]
      summary: Start a new backtest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestInput"
      responses:
        "202":
          description: Backtest started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backtest"
  /backtests/{backtestId}:
    get:
      operationId: getBacktest
      tags: [backtests]
      summary: Get backtest results
      parameters:
        - in: path
          name: backtestId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backtest results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backtest"
components:
  schemas:
...
    PnlDataPoint:
      type: object
      properties:
        timestamp:
          type: string
        pnl:
          type: number
        cumulativePnl:
          type: number
        trades:
          type: integer
      required: [timestamp, pnl, cumulativePnl, trades]
...
    Strategy:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [scalping, momentum, arbitrage, grid, sniping, ai_signal, manual, macd_cross, rsi_reversal, bb_squeeze]
        status:
          type: string
          enum: [active, inactive, error]
        exchangeId:
          type: string
          nullable: true
        symbols:
          type: array
          items:
            type: string
        config:
          type: object
          additionalProperties: true
        riskConfig:
          $ref: "#/components/schemas/RiskConfig"
        stats:
          $ref: "#/components/schemas/StrategyStats"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, description, type, status, exchangeId, symbols, config, riskConfig, stats, createdAt, updatedAt]
    StrategyList:
      type: object
      properties:
        strategies:
          type: array
          items:
            $ref: "#/components/schemas/Strategy"
        total:
          type: integer
      required: [strategies, total]
    RiskConfig:
      type: object
      properties:
        maxPositionSize:
          type: number
        maxOpenTrades:
          type: integer
        dailyLossLimit:
          type: number
        tradeСooldownSeconds:
          type: integer
        stopLossPercent:
          type: number
          nullable: true
        takeProfitPercent:
          type: number
          nullable: true
        trailingStopPercent:
          type: number
          nullable: true
      required: [maxPositionSize, maxOpenTrades, dailyLossLimit, tradeСooldownSeconds, stopLossPercent, takeProfitPercent, trailingStopPercent]
    StrategyStats:
      type: object
      properties:
        totalTrades:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        avgPnl:
          type: number
      required: [totalTrades, winRate, totalPnl, avgPnl]
    StrategyInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [scalping, momentum, arbitrage, grid, sniping, ai_signal, manual, macd_cross, rsi_reversal, bb_squeeze]
        exchangeId:
          type: string
          nullable: true
        symbols:
          type: array
          items:
            type: string
        config:
          type: object
          additionalProperties: true
        riskConfig:
          $ref: "#/components/schemas/RiskConfig"
      required: [name, type, symbols]
    StrategyUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        exchangeId:
          type: string
          nullable: true
        symbols:
          type: array
          items:
            type: string
        config:
          type: object
          additionalProperties: true
        riskConfig:
          $ref: "#/components/schemas/RiskConfig"

    # ── EXCHANGES ──────────────────────────────────────────────────────────────
    Exchange:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        exchangeId:
          type: string
        label:
          type: string
        status:
          type: string
          enum: [connected, disconnected, error, testing]
        isTestnet:
          type: boolean
        paperTrading:
          type: boolean
        rateLimit:
          type: integer
        createdAt:
          type: string
        lastConnected:
          type: string
          nullable: true
      required: [id, name, exchangeId, label, status, isTestnet, paperTrading, rateLimit, createdAt, lastConnected]
    ExchangeList:
      type: object
      properties:
        exchanges:
          type: array
          items:
            $ref: "#/components/schemas/Exchange"
        total:
          type: integer
      required: [exchanges, total]
    ExchangeInput:
      type: object
      properties:
        exchangeId:
          type: string
        label:
          type: string
        apiKey:
          type: string
        apiSecret:
          type: string
        isTestnet:
          type: boolean
          default: false
        paperTrading:
          type: boolean
          default: true
        passphrase:
          type: string
          nullable: true
      required: [exchangeId, label, apiKey, apiSecret]
    ExchangeTestResult:
      type: object
      properties:
        success:
          type: boolean
        latencyMs:
          type: number
        message:
          type: string
        permissions:
          type: array
          items:
            type: string
      required: [success, latencyMs, message, permissions]
    ExchangeBalance:
      type: object
      properties:
        exchangeId:
          type: string
        balances:
          type: array
          items:
            $ref: "#/components/schemas/AssetBalance"
        totalUsdValue:
          type: number
        fetchedAt:
          type: string
      required: [exchangeId, balances, totalUsdValue, fetchedAt]
    AssetBalance:
      type: object
      properties:
        asset:
          type: string
        free:
          type: number
        locked:
          type: number
        total:
          type: number
        usdValue:
          type: number
          nullable: true
      required: [asset, free, locked, total, usdValue]
...
    Backtest:
      type: object
      properties:
        id:
          type: string
        strategyType:
          type: string
        symbol:
          type: string
        exchange:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        config:
          type: object
          additionalProperties: true
        results:
          $ref: "#/components/schemas/BacktestResults"
          nullable: true
        createdAt:
          type: string
        completedAt:
          type: string
          nullable: true
      required: [id, strategyType, symbol, exchange, startDate, endDate, status, config, results, createdAt, completedAt]
    BacktestResults:
      type: object
      properties:
        totalTrades:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        totalPnlPercent:
          type: number
        maxDrawdown:
          type: number
        sharpeRatio:
          type: number
        profitFactor:
          type: number
        avgTradeMs:
          type: number
        equityCurve:
          type: array
          items:
            $ref: "#/components/schemas/PnlDataPoint"
      required: [totalTrades, winRate, totalPnl, totalPnlPercent, maxDrawdown, sharpeRatio, profitFactor, avgTradeMs, equityCurve]
    BacktestList:
      type: object
      properties:
        backtests:
          type: array
          items:
            $ref: "#/components/schemas/Backtest"
        total:
          type: integer
      required: [backtests, total]
    BacktestInput:
      type: object
      properties:
        strategyType:
          type: string
        symbol:
          type: string
        exchange:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        config:
          type: object
          additionalProperties: true
      required: [strategyType, symbol, exchange, startDate, endDate]