# lib/api-zod/src/generated/types/backtestList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { Backtest } from "./backtest";

export interface BacktestList {
  backtests: Backtest[];
  total: number;
}