# lib/api-zod/src/generated/types/exchangeBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetBalance } from "./assetBalance";

export interface ExchangeBalance {
  exchangeId: string;
  balances: AssetBalance[];
  totalUsdValue: number;
  fetchedAt: string;
}