# lib/api-zod/src/generated/types/exchangeTestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface ExchangeTestResult {
  success: boolean;
  latencyMs: number;
  message: string;
  permissions: string[];
}