# lib/api-zod/src/generated/types/strategyList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { Strategy } from "./strategy";

export interface StrategyList {
  strategies: Strategy[];
  total: number;
}